Preparation for next stable release
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/branches/stable@27341 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSFormat.m
@@ -1671,7 +1671,7 @@ NSDictionary *locale)
 	  {
 	    /* If the pointer is not NULL, write it as a %#x spec.  */
 	    base = 16;
-	    number.word = (unsigned long int) ptr;
+	    number.word = (size_t) ptr;
 	    is_negative = 0;
 	    alt = 1;
 	    group = 0;
